Fix Refactoring Structure Architecture
