--- app/src/main/java/com/vinylsmobile/activities/CollectionActivity.kt
@@ -1,31 +1,0 @@
-package com.vinylsmobile.activities
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.vinylsmobile.R
-import com.vinylsmobile.activities.ui.theme.VinylsMobileTheme
-import com.vinylsmobile.ui.fragments.AlbumFragment
-
-class CollectionActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_collection)
-
-        if (savedInstanceState == null) {
-            supportFragmentManager.beginTransaction()
-                .replace(R.id.fragment_container, AlbumFragment())
-                .commit()
-        }
-    }
-}
